requirements.txt
annotated-types
anyio
bcrypt
certifi
charset-normalizer
click
dnspython
email-validator
fastapi
h11
httpcore
httptools
httpx
idna
Jinja2
markdown-it-py
MarkupSafe
mdurl
passlib
pydantic
pydantic_core
PyJWT
python-dotenv
python-multipart
PyYAML
requests
sentry-sdk
shellingham
sniffio
SQLAlchemy
starlette
typer
typing-inspection
typing_extensions
urllib3
uvicorn
watchfiles
websockets
psycopg2-binary==2.9.8
